Backend/requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.1
astunparse==1.6.3
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
beautifulsoup4==4.9.3
black==19.10b0
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
Django==3.1.7
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
gspread==3.7.0
h5py==2.10.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.0.1
joblib==1.0.1
lxml==4.6.3
Markdown==3.3.4
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.23
nsepy==0.8
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.4
pathspec==0.7.0
Pillow==8.2.0
protobuf==3.15.8
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex @ file:///C:/ci/regex_1616096096031/work
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
termcolor==1.1.0
threadpoolctl==2.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
typed-ast @ file:///C:/ci/typed-ast_1610484654578/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.4
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.4.1
